Revert "Use a single reloader per tls config file, and not one per instance."
This reverts commit c58415566e23dcac5f0daa352f39f567a4d7b44f.
